tests/wpt/css-tests/css-writing-modes-3_dev/xhtml1/logicalprops-quirklength.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.1//EN" "http://www.w3.org/TR/xhtml11/DTD/xhtml11.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8" />
<title>CSS Logical Properties: {max-,min-}block-size</title>
<link href="mailto:openxu@gmail.com" rel="author" title="Xu Xing" />
<link href="https://drafts.csswg.org/css-logical-props-1/#logical-dimension-properties" rel="help" />
<link href="https://drafts.csswg.org/css-writing-modes-3/#logical-to-physical" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style> #dummy {} </style>

<script>
function isValidDeclaration(cssText) {
  var cssRule = document.styleSheets[0].cssRules[0];
  cssRule.style = cssText;
  var valid = (cssRule.style.length &gt; 0);
  cssRule.style = "";
  return valid;
}
var tests = [
  {cssText:"block-size: 1"},
  {cssText:"min-block-size: 1"},
  {cssText:"max-block-size: 1"},
  {cssText:"inline-size: 1"},
  {cssText:"min-inline-size: 1"},
  {cssText:"max-inline-size: 1"},
];
tests.forEach(function(t) {
  test(() =&gt; assert_false(isValidDeclaration(t.cssText)), "Check that unitless lengths are not valid for logical properties in quirks mode");
});
</script>
</head><body></body></html>
